Update tutorial.md
Mark, I'm still pretty new to programming culture so I don't know if this is trolling or not, but I assure you that is not my goal. I'm not very familiar with CoffeeScript... and the only way I know how to use backbonejs is the way Mason is teaching us, so I have a somewhat narrow understanding of both.

I made some changes inside the "Step 4:" section and I'm wondering how the way you did it would work? Are they typos, or just a different way of making backbone views? Are you going to talk about this at greenvillejs? if so, I can't wait!